Change how a bunch of stuff works.
Client is a bit higher level now: it has its own mirrored copy of the
game state (cursor positions and playing field state are what's
relevant). It does the rendering itself instead of relying on the
server, which will be important when we add filtering via the column
buttons (A-H).

Relatedly, the 'here are what lights to change on the launchpad'
approach has been replaced by an ad-hoc JSON thing (to be replaced with
an efficient serialization format later, if necessary), which gives the
state of the playing field and cursors, and lists any new events.

Events are what occur whenever any cursor hits something. Thus, even
simpler than the last revision of the idea, actual note triggers are
just effectors that do nothing at all. The mere fact of their being hit
will be used to tell the Max patch what to do.