Add warp effector, improve launchpad color mapping
"Warps warp the cursor to the next warp."
(The next warp is currently defined as the closest warp to the right.)
Assigned launchpad colors somewhat deliberately, and made the
filter/select buttons the same color as the associated effector. Also,
when you press them, they turn off to register the hold.